pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mastery-engine"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "rich>=13.0",
    "anthropic>=0.40",
    "openai>=1.0",
    "google-genai>=1.0",
    "google-api-python-client>=2.100",
    "google-auth>=2.20",
    "google-auth-oauthlib>=1.0",
]

[project.scripts]
mastery-engine = "mastery_engine.cli:main"
